Update Character
Character has been split into BaseCharacter, PlayerCharacter, and NPC.
Added ICharacter.
Added IPlayerCharacter.
Added Spawnable types.
Moved RegionActions to \Template\Types\
Added AIController placeholder.
Item price is now an int32.
Removed Currency from InventoryController.
Removed Currency from Character.
Removed Currency from Database CharacterEntity.
Currency is now a CharacterAttribute. This allows for easier additions of currency types.
Ability price is now an int32.

+++ FishMMO-Unity/Assets/Scripts/Shared/Entity/Ability/Template/BaseAbilityTemplate.cs
@@ -13,7 +13,7 @@ namespace FishMMO.Shared
 		public float Cooldown;
 		public float Range;
 		public float Speed;
-		public long Price;
+		public int Price;
 		public AbilityResourceDictionary Resources = new AbilityResourceDictionary();
 		public AbilityResourceDictionary Requirements = new AbilityResourceDictionary();
